requirements.txt
black==22.3.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
patsy==0.5.2
platformdirs==2.5.2
pyaml==21.10.1
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
scipy==1.8.1
six==1.16.0
statsmodels==0.13.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
